added add audit

+++ constants/firestoreManager.ts
@@ -18,8 +18,7 @@ export default class FirestoreManager<T extends DocumentData> {
     constructor(collectionName: string) {
         this.collectionName = collectionName;
     }
-
-    async create(data: T): Promise<string> {
+    async create(data: T | Record<string, any>): Promise<string> {
         try {
             const docRef = await addDoc(collection(db, this.collectionName), data);
             console.log(`${this.collectionName}: Belge başarıyla eklendi, ID: ${docRef.id}`);
@@ -29,7 +28,6 @@ export default class FirestoreManager<T extends DocumentData> {
             throw error;
         }
     }
-
     async getById(id: string): Promise<T | null> {
         try {
             const docRef = doc(db, this.collectionName, id);
@@ -75,8 +73,41 @@ export default class FirestoreManager<T extends DocumentData> {
     async delete(id: string): Promise<void> {
         try {
             const docRef = doc(db, this.collectionName, id);
+            const docSnap = await getDoc(docRef);
+    
+            if (!docSnap.exists()) {
+                console.warn(`${this.collectionName}: Belge zaten mevcut değil veya silinmiş.`);
+                return;
+            }
+    
             await deleteDoc(docRef);
-            console.log(`${this.collectionName}: Belge başarıyla silindi.`);
+            console.log(`${this.collectionName} id: ${id}: Belge başarıyla silindi.`);
+    
+            const verifySnap = await getDoc(docRef);
+            if (verifySnap.exists()) {
+                console.warn(`${this.collectionName}: Silme işlemi başarısız oldu.`);
+            } else {
+                console.log(`${this.collectionName}: Belge silme işlemi doğrulandı.`);
+            }
+        } catch (error) {
+            console.error(`${this.collectionName}: Belge silinirken hata oluştu:`, error);
+            throw error;
+        }
+    }
+    
+    async deleteByCustomId(customId: string): Promise<void> {
+        try {
+            const q = query(collection(db, this.collectionName), where("id", "==", customId));
+            const querySnapshot = await getDocs(q);
+    
+            if (querySnapshot.empty) {
+                console.warn(`${this.collectionName}: Belirtilen id (${customId}) ile eşleşen belge bulunamadı.`);
+                return;
+            }
+                const docRef = querySnapshot.docs[0].ref;
+            await deleteDoc(docRef);
+    
+            console.log(`${this.collectionName}: Custom ID (${customId}) ile belge başarıyla silindi.`);
         } catch (error) {
             console.error(`${this.collectionName}: Belge silinirken hata oluştu:`, error);
             throw error;
